+++ packages/flowr_dart/lib/src/mixin/loggable.dart
@@ -1,5 +1,4 @@
 import 'dart:developer' as dev;
-import 'dart:math';
 
 import 'package:flowr_dart/flowr_dart.dart';
 import 'package:stack_trace/stack_trace.dart';
@@ -7,63 +6,63 @@ import 'package:stack_trace/stack_trace.dart';
 ///
 ///   null: not print log extra info (dev tips, stack, ...)
 ///   [inner] : last FlowR method
-///   [self] : <dft> last your CustomViewModel(or other class) method
+///     - This is where you call the `::logger`
+///   [self] : (dft) last your CustomViewModel(or other class) method
+///     - This is where you call the `::update`
 ///   [outer] : invoke FlowR method at log.name
+///     - This is where you call the method that `contains the ::update` method
 ///   [all] : for dev, print all stack frame info
 enum LogExtra { inner, self, outer, all }
 
 /// 使用[logger] 打印异常信息
 mixin LoggableMx<T> {
+  static const _kExcludedPackages = {'flowr_dart', 'flowr', 'test_api'};
+
   ///
-  /// [logExtra] print stack frame info; (at log.name)
-  /// [uriFrame] show [logExtra] uri; (at log.message)
+  /// [logExtra] print stack frame info
+  /// [name] logger.name
+  ///   null: and if [logExtra] ==null: will use 'runtimeType'
+  ///         else: will use stack frame info
   logger(
     String message, {
     LogExtra? logExtra,
-    bool uriFrame = false,
     DateTime? time,
     int? sequenceNumber,
     int level = 0,
-    String? name, // null will use 'stateKey'
+    String? name,
     Zone? zone,
     Object? error,
     StackTrace? stackTrace,
+    @Deprecated('ignore this, always true') bool uriFrame = true,
   }) {
     if (logExtra != null) {
       try {
         final t = Trace.from(StackTrace.current);
-        // print('debug trace\n$t');
-        final maxAt = t.frames.length - 1;
-        int targetFrame = 0;
-        for (final t in t.frames) {
-          if (!'${t.uri}'.startsWith('package:flowr_dart/')) break;
-          targetFrame++;
-        }
-        final memberUriFm = switch (logExtra) {
-          LogExtra.inner => (-1, 0),
-          LogExtra.self => (0, 1),
-          LogExtra.outer => (0, 2),
-          LogExtra.all => (0, null),
-        };
-        final at = targetFrame + memberUriFm.$1;
-        name = t.frames[min(at, maxAt)].member;
-        //
-        if (uriFrame) {
-          final atOffset = memberUriFm.$2;
-          if (atOffset != null) {
-            final at = targetFrame + atOffset;
-            final fm = t.frames[min(at, maxAt)];
-            final location = fm.location;
-            final tips =
-                location.startsWith('package:flowr_dart/')
-                    ? '\t ----- DEV TIPS ----- \n'
-                        '\t Can not show correct location. You may need add "await" for VM::update method\n'
-                    : '';
-            // print('DEBUG: $message #> \n${t.frames.join('\n')}'); // fixme
-            message = '$message #> $location\n$tips';
-          } else {
-            message = '$message #> \n${t.frames.join('\n')}';
-          }
+        final fms0 = t.terse.frames;
+        final fms1 = fms0.where(
+          (f) =>
+              !_kExcludedPackages.contains(f.package) &&
+              // for `xx_test.dart`
+              (f.package != null || f.uri.scheme == 'file'),
+        );
+        // print('DEBUG fms1 ${fms1.join('\n\t')}');
+        if (fms1.isNotEmpty) {
+          name = fms1.first.member;
+          final fm = switch (logExtra) {
+            LogExtra.inner => fms1.firstOrNull,
+            LogExtra.self => fms1.skip(1).firstOrNull,
+            LogExtra.outer => fms1.lastOrNull,
+            LogExtra.all => null,
+          };
+          final locations =
+              fm?.location ?? fms1.map((f) => f.location).join('\n\t');
+          message = '$message #> $locations\n';
+        } else {
+          final tips =
+              '\t----- DEV TIPS:'
+              "\tCan't show correct invoke location. Try add 'await' for VM::update method\n"
+              '\t${fms0.join('\n\t')}\n';
+          message = '$message #> \n$tips';
         }
       } catch (e, s) {
         frPrint("FlowR LOGGER ERROR $e; \n$s");