+++ packages/flowr_dart/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.1.0 2025-8-31
+* fix LoggableMx::logger's LogExtra frames print; 
+* feat FlowR::put add logger;
+
 ## 2.0.0 2025-8-29
 * retract v1.3.0 , re publish as v2.0.0 
  - ::updateRaw -> ::update is break change